+++ src/main/java/com/application/close/links/controller/ModMqttLinksController.java
@@ -1,12 +1,16 @@
 package com.application.close.links.controller;
 
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.application.close.links.entity.ModMqttLinks;
 import com.application.close.links.service.ModMqttLinksService;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -29,11 +33,18 @@ public class ModMqttLinksController {
 		return ResponseEntity.ok("Linked Modbus TCP ID " + tcpId + " with MQTT Param ID " + paramId + " successfully.");
 	}
 
-	@Operation(summary = "Unlink Modbus TCP from MQTT Param", description = "Removes the existing mapping between a Modbus TCP client and an MQTT parameter.")
+	@Operation(summary = "Remove Modbus TCP–MQTT associations", description = "Unlinks Modbus TCP from all associated MQTT parameters and sets their values to zero")
 	@DeleteMapping("/disconnect/tcpId/{tcpId}")
 	public ResponseEntity<String> disconnectModFromMqtt(@PathVariable int tcpId) {
 		linksService.disconnectModFromMqtt(tcpId);
 		return ResponseEntity.ok("Unlinked Modbus TCP ID " + tcpId + " from its associated MQTT Param successfully.");
+	}
+
+	@Operation(summary = "Get all Modbus to MQTT links", description = "Retrieves all configured links between Modbus and MQTT")
+	@GetMapping
+	public ResponseEntity<List<ModMqttLinks>> getAll() {
+		List<ModMqttLinks> links = linksService.getAllLinks();
+		return ResponseEntity.ok(links);
 	}
 
 }