+++ src/main/java/com/application/close/mqtt/service/MqttService.java
@@ -1,5 +1,9 @@
 package com.application.close.mqtt.service;
 
+import java.util.List;
+
+import com.application.close.mqtt.entity.MqttParam;
+
 public interface MqttService {
 
 	String connect(int paramId); // Plain TCP connection
@@ -9,5 +13,7 @@ public interface MqttService {
 	String reconnect(int paramId);
 
 	void disconnect(int paramId);
+
+	List<MqttParam> getActiveMqtt();
 
 }